registry: add a way to configure passthrough registries
We want to allow users of the registry to be able to lazily move from
one registry to other and viceversa at some point.

Let's add a way to configure that.

Right now we expose it as a configuration that should be set under
[vars].

If empty or not set it should always be no-op for other users.

The schema is defined as a zod object like this:
 - "registry" should be the url of the target registry.
 - "password_env" should be the name of the environment variable that
   contains the registry password.
 - "username" should be a valid username to use on the target registry.

Related to issue #4.